index: add README.md and move README to index
When following the Github documentation for setting up a local build,
the current index.md simply shows the text "README.md".  I suspect this
is due to a version difference, and anyway the site shouldn't have the
same content as the repo.

Rename README.md to index.md, to keep the site with the same
functionality, and add a new README.md with some description.

Signed-off-by: Sean Cross <sean@xobs.io>